+++ zombsole/gym_env.py
@@ -4,6 +4,7 @@ from os import path, system
 from gym.core import Env
 from gym.spaces import Text, Box, Dict
 from gym.spaces.discrete import Discrete
+from gym.envs.registration import register
 from zombsole.gym.observation import build_observation
 from zombsole.gym.reward import AgentRewards
 from zombsole.game import Game, Map
@@ -384,4 +385,37 @@ class ZombsoleGymEnvDiscreteAction(Wrapper):
 
     def reverse_action(self, action):
         return self.game_actions.index(action)
+
+
+register(
+    id='jvstinian/Zombsole-v0', 
+    entry_point='zombsole.gym_env:ZombsoleGymEnvDiscreteAction', 
+    max_episode_steps=1000,
+    kwargs={
+        'rules_name': 'extermination',
+        'player_names': [],
+        'map_name': 'bridge',
+        'agent_id': 0,
+        'initial_zombies': 10,
+        'minimum_zombies': 0,
+        'debug': False
+    }
+)
+
+register(
+    id='jvstinian/Zombsole-SurroundingsView-v0', 
+    entry_point='zombsole.gym_env:ZombsoleGymEnvDiscreteAction', 
+    max_episode_steps=1000,
+    kwargs={
+        'rules_name': 'extermination',
+        'player_names': [],
+        'map_name': 'bridge',
+        'agent_id': 0,
+        'initial_zombies': 10,
+        'minimum_zombies': 0,
+        'observation_scope': 'surroundings:21',
+        'observation_position_encoding': 'simple',
+        'debug': False
+    }
+)